refactor: move utility classes to root namespace
- Relocate Util and FFT classes from Eduard.Cryptography to Eduard namespace
- Improve code organization and cross-module accessibility
- Streamline namespace hierarchy for cryptographic primitives
