+++ Eduard/FFT.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using Eduard;
 
-namespace Eduard.Cryptography
+namespace Eduard
 {
     public class FFT
     {
